+++ lib/events/new-editorial-comment.php
@@ -176,6 +176,29 @@ function comments_query( WP_Comment_Query $query ) {
 add_action( 'pre_get_comments', __NAMESPACE__ . '\comments_query' );
 
 /**
+ * Filter the comment count to exclude workflow comments.
+ *
+ * @param int $count The current number of comments found.
+ * @param int $post_id The post ID to return the number of comments for.
+ * @return int
+ */
+function exclude_workflow_comments_from_count( $count, $post_id ) {
+	if ( intval( $count ) === 0 ) {
+		return $count;
+	}
+
+	$workflow_comment_count = get_comments( [
+		'post_id' => $post_id,
+		'type__in' => [ 'workflow' ],
+		'count' => true,
+	] );
+
+	return $count - $workflow_comment_count;
+}
+
+add_filter( 'get_comments_number', __NAMESPACE__ . '\exclude_workflow_comments_from_count', 10, 2 );
+
+/**
  * Register assignees meta.
  */
 function assignees_api() {